# src/main/java/com/auction/system/controller/UsersApiDelegateImpl.java
package com.auction.system.controller;

import java.util.List;

import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.auction.system.generated.controllers.UsersApiDelegate;
import com.auction.system.generated.models.UserDto;
import com.auction.system.service.UserService;

import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("api/v1/users")
@RequiredArgsConstructor
public class UsersApiDelegateImpl implements UsersApiDelegate {

  private final UserService userService;

  @Override
  public ResponseEntity<UserDto> getCurrentUserProfile() {
    return ResponseEntity.ok(userService.getProfile());
  }

  @Override
  @PreAuthorize("hasRole('ADMIN')")
  public ResponseEntity<List<UserDto>> getAllUsers() {
    return ResponseEntity.ok(userService.getAllUsers());
  }

}