# src/main/resources/api/openapi.api-description.yaml
openapi: 3.0.3
info:
  title: Online Auction System
  version: 0.5.1
  description: |
    This is an OpenAPI description of an **Online Auction System**.
servers:
  - url: http://127.0.0.1:8080/api/v1
...
paths:
...
  /users/me:
    get:
      description: |
        Retrieves the logged-in user's profile.
      summary: Retrieves the logged-in user's profile.
      operationId: getCurrentUserProfile
      tags:
        - Users
      responses:
        "200":
          description: Successfully retrieved user's profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/DefaultError"
...
components:
  schemas:
    User:
      description: User DTO
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
          pattern: "^[A-Za-z0-9_-]+$"
          maxLength: 16
        email:
          type: string
          format: email
        contactNumber:
          type: string
        shippingAddr:
          $ref: "#/components/schemas/UserAddress"
    UserAddress:
      description: User address.
      type: object
      properties:
        row1:
          type: string
        row2:
          type: string
        row3:
          type: string
...
    ApiError:
      description: |
        Central error response object.
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    DefaultError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Authorization using JWT issued by the backend.
      bearerFormat: JWT
    oauth2google:
      type: http
      scheme: bearer
      description: Authentication using ID Tokens from the Google Authorization Server.
      bearerFormat: ID Token (Google)
    oauth2apple:
      type: http
      scheme: bearer
      description: Authentication using ID Tokens from the Apple Authorization Server.
      bearerFormat: ID Token (Apple)
security:
  - bearerAuth: []